// src/types.ts
export type CourseDate = {
  id: string;
  date: string;
  time: string;
  venue: string;
  isActive: boolean;
};

export type FormData = {
  email: string;
  date: string;
  name: string;
  participants: string;
  phone: string;
  applicationNumber: string;
  experience: string;
  questions: string;
  privacy: boolean;
};